1、修改了RpcMessage中requestId的类型 2、实现了编码解码 3、添加了一些常量

+++ rpc-framework-core/src/main/java/github/javabro/remoting/dto/RpcMessage.java
@@ -24,7 +24,7 @@ public class RpcMessage {
     //压缩类型
     private byte compress;
     //请求的id
-    private byte request;
+    private int request;
     //数据
     private Object data;
